Big refactoring: compiler warnings and errors

+++ deps/uv/src/unix/getaddrinfo.c
@@ -106,7 +106,7 @@ static void uv__getaddrinfo_work_wrapper (uv_work_t *req)
 {
   uv_getaddrinfo_t *addr_req;
   addr_req = (uv_getaddrinfo_t *) req->data;
-  INVOKE_CALLBACK_1(UV_GETADDRINFO_WORK_CB, uv__getaddrinfo_work, &addr_req->work_req);
+  INVOKE_CALLBACK_1(UV_GETADDRINFO_WORK_CB, uv__getaddrinfo_work, (long) &addr_req->work_req);
 }
 
 void * uv_uv__getaddrinfo_work_wrapper_ptr (void)
@@ -147,7 +147,7 @@ static void uv__getaddrinfo_done(struct uv__work* w, int status) {
   if (req->cb)
   {
 #if UNIFIED_CALLBACK
-    INVOKE_CALLBACK_3(UV_GETADDRINFO_CB, req->cb, req, req->retcode, req->addrinfo);
+    INVOKE_CALLBACK_3(UV_GETADDRINFO_CB, req->cb, (long) req, (long) req->retcode, (long) req->addrinfo);
 #else
     req->cb(req, req->retcode, req->addrinfo);
 #endif
